+++ LooperFixTool/Program.vb
@@ -1,0 +1,7 @@
+Module Program
+    Sub Main()
+        Application.EnableVisualStyles()
+        Application.SetCompatibleTextRenderingDefault(False)
+        Application.Run(New MainForm())
+    End Sub
+End Module